refactor(simulation): split test into modular structure
Split the large simulate_replication.rs file into organized modules:
- config.rs: Constants and simulation configuration
- counters.rs: Global atomic counters for statistics
- types.rs: Core data structures and ReplicationAccumulator
- node.rs: SimulatedNode implementation
- stats.rs: UploadStats and CoverageStats helpers
- export.rs: JSON export structures
- mod.rs: Main test function and report generation
